Refactored infrastructure
Removed some huge size tarballs

Installed dependencies online

Updated generate.py

Splited Dockerfile.kerneldev to Dockerfile.dependency

+++ generate.py
@@ -1,13 +1,17 @@
-import jinja2
 from jinja2 import Template
+from yaml import safe_load
 import os
 import shutil
 import time
+import sys
+import getopt
+import platform
 
 class Assembler(object):
     def __init__(self):
         self.devtool = 'Dockerfiles/Dockerfile.devtool'
         self.kerneld = 'Dockerfiles/Dockerfile.kerneldev'
+        self.dependf = 'Dockerfiles/Dockerfile.dependency'
         self.current = os.getcwd()
         self.dt_path = os.path.join(self.current, self.devtool)
         self.kd_path = os.path.join(self.current, self.kerneld)
@@ -15,6 +19,7 @@ class Assembler(object):
         self.tempdir = os.path.join(self.current, 'Dockerfiles/temp')
         self.readbuf = ''
         self.writbuf = ''
+        self.signatu = 'clayding <gdskclay@gmail.com>'
 
 
     def read(self, file):
@@ -32,8 +37,17 @@ class Assembler(object):
     def assemble(self, src, dst, new):
         srcbuff = self.read(src)
         dstbuff = self.read(dst)
+
+        distinfo = platform.dist()
+        print(self.get_dist_version(distinfo[0]))
+        base_ver = distinfo[0] + ':' + str(self.get_dist_version(distinfo[0]))
+        base_dep = self.get_dist_dependency(distinfo[0])
+
         template= Template(dstbuff)
-        content = template.render(DOCKFILE_DEVTOOL=srcbuff)
+        content = template.render(DOCKERFILE_DEVTOOL=srcbuff,
+                                  DOCKERFILE_OS_VERSION=base_ver,
+                                  DOCKERFILE_DEPENDENCIES=base_dep,
+                                  DOCKERFILE_MAINTAINER=self.signatu)
         #print(content)
 
         self.write(new, content)
@@ -49,11 +63,12 @@ class Assembler(object):
     
     def create_link(self, path):
         rootpath = self.dockerf
-        isexist = os.path.exists(rootpath)
-        if isexist:
+        if os.path.exists(rootpath):
             os.unlink(rootpath)
-        os.symlink(path, rootpath)
-        print('Create new symbol link to:{}' .format(path))
+
+        if os.path.exists(path):
+            os.symlink(path, rootpath)
+            print('Create new symbol link to:{}' .format(path))
     
     def clean(self):
         rootpath = self.dockerf
@@ -76,11 +91,43 @@ class Assembler(object):
         self.assemble(self.dt_path, self.kd_path, newpath)
         self.create_link(newpath)
 
-import sys, getopt
+    def read_yaml(self, filename):
+        with open(filename, 'r') as f:
+            buffer = f.read()
+        f.close()
+        vars = safe_load(buffer)
+        return vars
+
+    def get_dist_version(self, dist):
+        yamlvars = self.read_yaml(self.dependf)
+        distinfo = yamlvars[dist]
+        if distinfo:
+            return distinfo['version']
+        else:
+            print("Not support {}" .format(dist))
+            return NULL
+    
+    def get_dist_dependency(self, dist):
+        yamlvars = self.read_yaml(self.dependf)
+        distinfo = yamlvars[dist]
+        if distinfo:
+            return distinfo['dependency']
+        else:
+            print("Not support {}" .format(dist))
+            return NULL
+
+    def list_dist(self):
+        yamlvars = self.read_yaml(self.dependf)
+        print('List suported distributions: ', end = '')
+        for key in yamlvars:
+            print('{} ' .format(key), end = '')
+        print('')
+        print('Current platform distribution: {}' .format(platform.dist()))
 
 def parse_options(argv):
     exename = argv[0]
     ab = Assembler()
+    ab.list_dist();
     try:
         opts, args = getopt.getopt(argv[1:], "hck", ["kernel=", "clean="])
     except getopt.GetoptError:
